Use int_least8_t for rise and run

+++ src/shadowcasting.cpp
@@ -7,7 +7,7 @@
 #include "line.h"
 
 struct slope {
-    slope( int rise, int run ) {
+    slope( int_least8_t rise, int_least8_t run ) {
         // Ensure run is always positive for the comparison operators
         this->run = abs( run );
         if( run < 0 ) {
@@ -17,8 +17,10 @@ struct slope {
         }
     }
 
-    int rise;
-    int run;
+    // We don't need more that 8 bits since the shadowcasting area is not that large,
+    // currently the radius is 60.
+    int_least8_t rise;
+    int_least8_t run;
 };
 
 inline bool operator<( const slope &lhs, const slope &rhs )
